Adding JWT access token authentication.

+++ resources/item.py
@@ -1,5 +1,6 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 
 from db import db
@@ -9,26 +10,6 @@ from schemas import ItemSchema, ItemUpdateSchema
 blp = Blueprint("items", __name__, description="Item-Related Actions")
 
 
-@blp.route("/item")
-class Items(MethodView):
-    @blp.response(200, ItemSchema(many=True))
-    def get(self):
-        return ItemModel.query.all()
-
-    @blp.arguments(ItemSchema)
-    @blp.response(201, ItemSchema)
-    def post(self, item_data):
-        item = ItemModel(**item_data)
-
-        try:
-            db.session.add(item)
-            db.session.commit()
-        except SQLAlchemyError:
-            abort(500, message="An error ocurred while trying to create the item.")
-
-        return item
-
-
 @blp.route("/item/<int:item_id>")
 class Item(MethodView):
     @blp.response(200, ItemSchema)
@@ -36,6 +17,7 @@ class Item(MethodView):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required()
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
     def put(self, item_data, item_id):
@@ -51,8 +33,38 @@ class Item(MethodView):
         db.session.commit()
         return item
 
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+
+        if not jwt.get("is_admin"):
+            abort(
+                401,
+                message="Administrator priviliges are required to perform this operation.",
+            )
+
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
         return {"status": 200, "message": "The item has been deleted."}
+
+
+@blp.route("/item")
+class Items(MethodView):
+    @blp.response(200, ItemSchema(many=True))
+    def get(self):
+        return ItemModel.query.all()
+
+    @jwt_required()
+    @blp.arguments(ItemSchema)
+    @blp.response(201, ItemSchema)
+    def post(self, item_data):
+        item = ItemModel(**item_data)
+
+        try:
+            db.session.add(item)
+            db.session.commit()
+        except SQLAlchemyError:
+            abort(500, message="An error ocurred while trying to create the item.")
+
+        return item
