+++ resources/user.py
@@ -1,13 +1,39 @@
 from flask.views import MethodView
+from flask_jwt_extended import create_access_token, get_jwt, jwt_required
 from flask_smorest import Blueprint, abort
 from passlib.hash import pbkdf2_sha256
 from sqlalchemy.exc import SQLAlchemyError
 
+from blocklist import BLOCKLIST
 from db import db
 from models import UserModel
 from schemas import UserSchema
 
 blp = Blueprint("users", __name__, description="User-Related Actions")
+
+
+@blp.route("/login")
+class UserLogin(MethodView):
+    @blp.arguments(UserSchema)
+    def post(self, user_data):
+        user = UserModel.query.filter(
+            UserModel.username == user_data["username"]
+        ).first()
+
+        if user and pbkdf2_sha256.verify(user_data["password"], user.password):
+            access_token = create_access_token(identity=str(user.id))
+            return {"access_token": access_token}
+
+        abort(401, message="The credentials are invalid.")
+
+
+@blp.route("/logout")
+class UserLogout(MethodView):
+    @jwt_required()
+    def post(self):
+        jti = get_jwt()["jti"]
+        BLOCKLIST.add(jti)
+        return {"status": 200, "message": "You have successfully logged out."}
 
 
 @blp.route("/register")
@@ -39,7 +65,16 @@ class User(MethodView):
         user = UserModel.query.get_or_404(user_id)
         return user
 
+    @jwt_required()
     def delete(self, user_id):
+        jwt = get_jwt()
+
+        if not jwt.get("is_admin"):
+            abort(
+                401,
+                message="Administrator priviliges are required to perform this operation.",
+            )
+
         user = UserModel.query.get_or_404(user_id)
         db.session.delete(user)
         db.session.commit()
